backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.1.0
pymysql==1.1.1
mysql-connector-python